+++ src/main/java/net/pwall/pipeline/AbstractIntPipeline.java
@@ -2,7 +2,7 @@
  * @(#) AbstractIntPipeline.java
  *
  * pipelines   Pipeline conversion library for Java
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -44,6 +44,11 @@ public abstract class AbstractIntPipeline<R> extends AbstractIntAcceptor<R> impl
         this.downstream = downstream;
     }
 
+    @Override
+    public IntAcceptor<? extends R> getDownstream() {
+        return downstream;
+    }
+
     /**
      * Close the pipeline.
      */
@@ -61,27 +66,6 @@ public abstract class AbstractIntPipeline<R> extends AbstractIntAcceptor<R> impl
     @Override
     public void emit(int value) throws Exception {
         downstream.accept(value);
-    }
-
-    /**
-     * Get the result object (defaults to the result of the downstream acceptor).
-     *
-     * @return  the result
-     */
-    @Override
-    public R getResult() {
-        return downstream.getResult();
-    }
-
-    /**
-     * Return {@code true} if all sequences in the input are complete, that is, the input is not in the middle of a
-     * sequence requiring more data.  The default implementation tests whether the downstream acceptor is complete.
-     *
-     * @return  {@code true} if the input is in the "complete" state
-     */
-    @Override
-    public boolean isComplete() {
-        return downstream.isComplete();
     }
 
     /**