+++ src/main/java/net/pwall/pipeline/IntPipeline.java
@@ -2,7 +2,7 @@
  * @(#) IntPipeline.java
  *
  * pipelines   Pipeline conversion library for Java
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -32,7 +32,7 @@ package net.pwall.pipeline;
  * @author  Peter Wall
  * @param   <R>     the result type
  */
-public interface IntPipeline<R> extends IntAcceptor<R> {
+public interface IntPipeline<R> extends IntAcceptor<R>, BasePipeline<R> {
 
     /**
      * Emit a value, that is, forward a value to the downstream acceptor.