+++ src/main/java/net/pwall/pipeline/codec/UTF8_CodePoint.java
@@ -2,7 +2,7 @@
  * @(#) UTF8_CodePoint.java
  *
  * pipelines   Pipeline conversion library for Java
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -76,8 +76,8 @@ public class UTF8_CodePoint<R> extends AbstractIntPipeline<R> {
     }
 
     @Override
-    public boolean isComplete() {
-        return state == normal && super.isComplete();
+    public boolean isStageComplete() {
+        return state == normal;
     }
 
     private void startSequence(int i, IntConsumer nextState) {